fixed so creation table is generic and not needed to be done every time
