+++ src/routes/admin/news/News.tsx
@@ -9,6 +9,7 @@ import {
 	getPaginationRowModel,
 } from "@tanstack/react-table";
 import AdminTable from "@/widgets/AdminTable";
+import createTable from "@/widgets/createTable";
 const ACCEPTED_IMAGE_TYPES = [
 	"image/jpeg",
 	"image/jpg",
@@ -26,17 +27,13 @@ export interface NewsItem {
 const columnHelper = createColumnHelper<NewsRead>();
 
 const columns = [
-	columnHelper.accessor((row) => row.title_sv, {
-		id: "title_sv",
+	columnHelper.accessor("title_sv", {
+		header: "Svensk titel",
 		cell: (info) => info.getValue(),
-		header: () => <span>Svensk titel</span>,
-		//footer: (props) => props.column.id,
 	}),
-	columnHelper.accessor((row) => row.content_sv, {
-		id: "content_sv",
+	columnHelper.accessor("content_sv", {
+		header: "Svensk beskrivning",
 		cell: (info) => info.getValue(),
-		header: () => <span>Svensk beskrivning</span>,
-		//footer: (props) => props.column.id,
 	}),
 ];
 
@@ -47,18 +44,7 @@ export default function News() {
 		...getAllNewsOptions(),
 	});
 
-	const table = useReactTable({
-		columns,
-		data: (data as NewsRead[]) ?? [],
-		getCoreRowModel: getCoreRowModel(),
-		getPaginationRowModel: getPaginationRowModel(),
-		initialState: {
-			pagination: {
-				pageIndex: 0,
-				pageSize: 10,
-			},
-		},
-	});
+	const table = createTable({ data: data ?? [], columns });
 
 	if (isFetching) {
 		return <p> Hämtar</p>;
